fix(services): wait for all promises involving user details
The way the user details were being fetched was via a users.map() loop, which ended up defining an async function inside. Because of this, the details were actually returning a promise inside the map, which could not be understood by the application, so I needed to find a way to resolve them.

Looking up online, I've found two ways: (1) keep the users.map() call returning promises, then call Promise.all(...) to process them all at the same time, or (2) change the map() loop to a for .. of loop which keeps the async/await right outside it. Although (2) was probably more readable (although for .. of isn't common in JavaScript, to me at least), the usage of await meant that all requests needed to be waited synchronously, whereas Promise.all(...) was capable to wait for all of them while keeping the async fashion.

I ended up going with (1), and used Promise.all() to process all user details before returning to fetchMore().