fix(models): rename public_gists to gists which is expected onscreen
I found it weird that the gist count, expected on Figma, was not appearing for _any_ user at all. By accident, I noticed that the user model was expecting the data to contain gists instead of public_gists, so I've converted that in the User model, which was used to map data in fetchMore().

As soon as I did that, the first test started to pass.