fix(scroll): do not trigger more actions when one is ongoing
Part 1 of the bugfix for the problem of triggering multiple actions to fetch a list of GitHub users.

Although a reducer is a good place to handle immutable states, the state changes are handled purely and in async form, which means a "duplicate" action will affect the state twice. The initial version of the infinite scroll, by working with an inequality in regards to scroll offsets, could potentially trigger 2+ scroll events satisfying the inequality, effectively fetching the same user list multiple times in one scroll to the bottom.

Similar to a [semaphore](https://en.wikipedia.org/wiki/Semaphore_(programming)), I was able to use the loading state triggered by the FETCH action to tell whether we were fetching users at a given time during an infinite scroll, and avoid calling fetchMore() unnecessary times. I did consider using this inside fetchMore() instead of useInfiniteScroll(), but only this hook had such a problem.