fix(scroll): use useState to avoid triggering duplicate scroll actions
This is Part 2 of the multiple user fetch bugfix.

Part 1 involved using the loading state of the useReducer to avoid triggering multiple fetchMore() actions. However, as we were working with an inequality in order to reach the desired edgeOffset (800), until the loading state changed to block more fetchMore() calls, there was a small window of time where another scroll event satisfying the inequality could happen, triggering another fetchMore() call.

There was the possibility of ditching the edgeOffset and triggering the infinite scroll only when the user reached the bottom of the grid, which would not bring such a window of opportunity for a user to re-scroll and trigger another fetchMore(). However, the edgeOffset was part of the requirements in the README, so I tried working with it as much as possible, which basically meant I needed to use another [semaphore](https://en.wikipedia.org/wiki/Semaphore_(programming)) and make another section of the code thread-safe.

This time, I opted to use a simple useState() hook storing a boolean. Right before the infinite scroll triggered an action, the semaphore would be switched to true. If any scroll event happened in succession, a 'true' semaphore would block it. When the action was completed, the semaphore would be switched to false, allowing for another event to trigger it.

Funny enough, I was able to trigger duplicate fetchMore() calls whenever I tried using only Part 1 or Part 2, but not with both.