fix(tests): fix all failing tests
The first test, listing the first 50 users as soon as you load the page, solved itself when I renamed public_gists to gists in the model function.

When trying to load the remaining tests, it was weird that the 52nd user never loaded, so I started logging the reducer in the Assignment component and even started to run each of the remaining three tests individually using it.only. What I soon figured out was that tests were easily timing out for no apparent reason, as the next 50 users were being fetched. I then guessed that the extra waitForDomChange (with a sneaky comment to induce safety) was the culprit, so I removed it. As soon as I did that, the second test started to pass.

I then adjusted the other two tests to properly scroll and present the 52nd user. They kept failing, however, as they were relying on mocking the window.location.assign() function while the code was using window.location.href, which is functionally similar but complicates testing, so I've changed the code to start using window.location.assign().

All tests are passing, now.