fix(css): use rem instead of px whenever possible
This might be more of a nitpicking of mine, but whenever I see part of the code using rem for measurements, I tend to assume the code was structured to use rem, and in this case, the entire code should be using rem, as the application becomes far more adaptable to be resized when using it.

I avoided applying it in @media-related breakpoints and shadows. There may be a few more places to consider applying (or not) rem to the measurements, but I would then prefer to discuss with the rest of the team about it.