feat(Assignment): implement search over filtered users
The search goes over each of the users' name and login. The search input is stored in a state, but the filter is done dynamically when the users are loaded into the `<Assignment />`. As the pagination could go in the way of the search (imagine scrolling to the end of the page, and nothing new appears onscreen because of the filter, but we keep hitting GitHub's API), any search input blocks the infinite scroll.